requirements.txt
camelot-py[cv]
opencv-python
pdf2image
ghostscript
pandas